Pass user options and worker context to reduce(), aggregate() and forEach() callbacks.
Despite being documented in the API reference, optional user defined argument and worker
context were not passed to reducer and combiner callbacks. This is now fixed.
The number of workers in local mode is no longer hard-coded, and can be set by SKALE_WORKERS
env variable, or the number of CPUs - 1 (if more than 2 CPUs).