Further improvements in serialization of tasks.
In datasets, convert all methods to prototypes rather than inlined in
objects, so the code of methods is not included in serialized objects.
Methods prototypes are then added manually to task in parseTask.
In context-local, if serialized objects is big, write it to a gzipped
file and send the filename to worker instead of sending a big request.