test: start to add tape test cases, as per #205
Skale context is shared amongst individual test files, so init is
quite fast. Each individual test case is self contained and can be
executed separately, which is good for debugging.

This is a work in progress, more tests need to be added / migrated.
Tape based tests must be run using `npm run tape`.