Implement dependency injection into workers, resolves #203
A new skale context method `sc.require` is added. It specifies a
set of modules on user side (master) which have to be deployed in
workers for use by callbacks, such as mappers, reducers, etc.

Under the hood, [browserify] is used on master side to build a
bundle which is serialized and sent to workers (as part of task).
It is then evaluated in worker global context, and dependencies
remain persistent as long as workers live.

This method allows to use in workers any javascript module which
can be browserified, so a large number (almost any pure JS package).

The current commit is for local version, not distributed (code will
be exactly the same). It is experimental for the moment.

An additional statement `sc.bundle` could be added as well to inject
pre-compiled modules, avoiding the penalty of `browserify` at each
run.

[browserify]: https://github.com/browserify/browserify